requirements.txt
# Groq (free tier LLM — 6000 req/day)
groq>=0.11.0

# Agent Framework
langchain>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.40

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0

# UI
streamlit>=1.30.0

# Web Search (free, no API key needed)
ddgs>=1.0.0

# Optional (for document search)
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
